fix(cli): harden localized notes morph and validation messages
Share one null-prototype locale-notes schema that rejects blank/duplicate
keys, keep paths on ArkType problems, and throw a typed error for
unsupported JSON Schema units.
