+++ cli/src/schemas/build.ts
@@ -1,21 +1,6 @@
 import { type } from './arktype'
+import { localizedReleaseNotesSchema } from './common'
 import { optionsBaseSchema } from './base'
-
-const localizedReleaseNotesSchema = type({ '[string]': 'string' }).pipe((data, ctx) => {
-  const out: Record<string, string> = {}
-  for (const [rawKey, rawValue] of Object.entries(data)) {
-    const key = rawKey.trim()
-    const value = rawValue.trim()
-    if (!key) {
-      return ctx.reject('a non-empty locale key')
-    }
-    if (!value) {
-      return ctx.reject('a non-empty release note')
-    }
-    out[key] = value
-  }
-  return out
-})
 
 // ============================================================================
 // Build Credentials Schema
